Remove Plot.plotSize and instead provide the size to render at render-time.
The idea of Plots being able to decide their own size was always a bit of a lie - if you initialised a BarGraph with a size, then put it in a SubPlot, its size would be overridden anyway.
Removing this has lots of advantages when integrating with native views, because it means they (via the Renderer) control their size as they always have.