docs: update dependency addition (#61)
The 1.0.0 release has a separate dependency on CFreeType that seems to have a problem. A recent PR moved that dependency into the package itself, as SwiftPM now supports system library targets. This solved the issue. SwiftPlot uses Anti-Grain Geometry (AGG), a C++ library as a rendering backend, which in turn uses freetype to render text.