bfix: "fix" bug where in some cases and some renderers the x labels showed values with too many decimal places due to numeric error.
In reality this just decreases the possibility of this happening again by reducing an extra division followed by a multiplication by the same value, which in some cases would result in loss of precision to the point where the string representation also had to change and show more decimal places.
We should in the future round the number to a certain number of decimal places by default and give the user control over it.