[bfix][rfac]: Fix bug where white lines appear through histogram bin edges due to previous commit.
This is the easy approach, in the case of multiple series the colour of the series in the back might/will bleed through the edges of the one in the front, even when/if there are no columns of the back series nearby.
Also refactored the code inside `drawData` slightly for better code reuse between `.bar` and `.step` cases.