bfix: Colour from some series in the background were bleeding through some edges due to the last commit, this fixes that with an improved algorithm.
This is the same idea as with the `.step` case, using conditions and rules to decide what points to add to form a polygon. In this case "tucking" the background series a little behind the foreground one by `floor`ing or `ceil`ing the right points.
Also improved the algorithm for `.step` by removing one "instruction" and reducing the cases some of the instructions were being emitted redundantly to the minimum possible, while preserving the same output.